requirements.txt
langgraph
langchain
langchain-google-genai
langchain-mcp-adapters 
anyio 
python-dotenv
networkx 
matplotlib
line-bot-sdk
geopy
fastapi
uvicorn
pymongo
mcp
notebook